Parsing script for use with future Drupal Uploads
Parser.py and header.py are both still in development and will depend on the exact Drupal format that is chosen. Currently the  most significant issue is:
Script fails if non-numerical field present in the actual data (this happens for instance when the x-axis contains names of different experiments). As a side effect, this means that unpopulated fields also result in parsing issues.